allow to change image size via HTTP headers, too
If the request contains the headers X-Image-Width or
X-Image-Height, then the PNG file's size will be adjusted, too.
Acceptable values for those headers are integers, e. g.:

    X-Image-Width: 900
    X-Image-Height: 500

If the headers do not contain a parseable integer, then they will
be ignored. Fallback will be the imageHeight and imageWidth data
members of the POST-ed JSON. If they are not integers or are not
present, then the fallback values will be used. Fallback for the
image width is 600 pixels, fallback for the image height is 400
pixels. This is the size that was used before this feature has
been introduced.