use Node.js 16 instead of Node.js 14 in Dockerfile
The previously used Node.js 14 is an old LTS version that has
reached its end of support in April 2023, so it should no longer
be used.

I would even go further and update to Node.js 18 or Node.js 20,
but the Dockerfile job on GitHub Actions fails for a yet unknown
reason when the container has to start with Node.js 18 or 20:

    curl: (7) Failed to connect to 127.0.0.1 port 3000 after 0 ms: Connection refused

Needs more investigating, but for now a smaller update that works
(Node.js 16) is still better than no update at all.