Revert "use Node.js 22 instead of Node.js 20 in Dockerfile"
This reverts commit f0e0dd4994b2ed560b2c4d84316e27065f171cdf.

Currently, there is no prebuilt for the canvas package for
Node.js 22, so we would have to build it from source and install
all the required dependencies. That would increase the size of
the final Docker image to almost twice the current size, and I
don't think that is acceptable. So let's wait for a newer release
of canvas which brings the corresponding prebuilt package.