documentation: update installation instructions to use Node.js 20
The previously mentioned Node.js 16 is no longer supported by the
Node.js project and does not receive any maintenance or updates
anymore. In other words: It has reached its end of life and is no
longer supported. Therefore the instructions should switch to a
Node.js version that is still maintained.

[ci skip]